refactor(tests): consolidate Runtime Artifact Layout Module — 12 files → 3
Closes #3757
